restructure file directories
